chore(benchmarks): rename module
